backend/requirements.txt
Django==5.0.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
django-ratelimit==4.1.0
celery==5.4.0
redis==5.0.7
django-redis==5.4.0
psycopg2-binary==2.9.9
cryptography==42.0.8
python-dotenv==1.0.1
resend==2.3.0
requests==2.32.3
python-docx==1.1.2
openpyxl==3.1.5
reportlab==4.2.2
Pillow==10.4.0
django-cors-headers==4.4.0
gunicorn==22.0.0
whitenoise==6.7.0
celery[redis]==5.4.0
